+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.9.11 (2019-02-12)
 
 * Set property values on all related mocks #944
+* Backported from master: Added possibility to add Constructor-Expectations on hard dependencies, read: Mockery::mock('overload:...')
 
 ## 0.9.4 (2015-04-02)